fix(test): the web project could not load its own database tests
`vitest.web.config.ts` aliased `@rostr/db` and `@rostr/db/postgres` but not
`@rostr/db/testing`, so `visibility.test.ts` failed to collect under
`pnpm test:web` — "Cannot find module '@rostr/db/testing'", zero tests run —
while passing under `pnpm test`, whose config has the entry.

Its six tests were not being skipped. They were never loading. They cover
whether a private league is actually private, which is a guarantee members read
in the frozen rule set, and `pnpm test:web` is the only project that runs them
against a real Postgres rather than skipping on a missing `DATABASE_URL`. A test
that passes in one project and cannot load in another is worse than a failing
one, because the green run is the one people quote.

**The ordering rule that fixes it was already written down** — in
`vitest.program.config.ts`'s own comment, "vite prefix-matches, so
`@rostr/db/testing` has to come before `@rostr/db` or it would resolve to the
main entry" — and had simply never reached the web config. Three configs held
three hand-maintained copies of the same list, and the web one had drifted
furthest: no `@rostr/db/testing`, no `@rostr/db/migrate`, no `@rostr/pinning`,
no `@rostr/stats`, and a `"@/"` string entry where the main project uses a regex
that matches a different set of specifiers.

So this adds one list rather than one entry. Copying the missing line across
would leave three lists to keep in step and the next package one omission away
from the same bug. `vitest.alias.ts` exports `workspaceAliases` (subpaths first,
with the ordering rule stated where it can be read before editing) and
`webAliases`, and all three projects take the whole list — a package a project
never imports costs nothing, and curating the subsets is what caused this.

Also removes `apps/web/test/server-only.ts`, an undocumented second stub
containing `export {}`. `test-stubs/server-only.ts` says what the marker is for
and why stubbing beats deleting the import; the duplicate said nothing and only
one project used it.

`pnpm test:web` now runs 60 tests, up from 54 — the difference is
`visibility.test.ts` executing for the first time. All six pass.

Negative control: deleting the `@rostr/db/testing` line reproduces the original
failure exactly, `0 test` and the same message.

`pnpm test` unchanged at 1300 passed, 2 skipped. Typecheck, lint, format clean.
`vitest.program.config.ts` is in `anchor.yml`'s trigger paths, so CI runs the
program suite against this.
