--- apps/web/test/server-only.ts
@@ -1,1 +1,0 @@
-export {};
